requirements.txt
fastapi~=0.115.11
uvicorn~=0.34.0
sqlalchemy
pydantic~=2.10.6
asyncpg
alembic
python-dotenv
passlib~=1.7.4
pyjwt~=2.10.1

pydantic-settings~=2.8.1